# contracts/http/openapi.yaml
openapi: 3.1.0
info:
  title: vlxd-v3 API
  version: 0.1.0
  description: Contract for the vlxd-v3 construction-materials management API.
servers:
  - url: http://localhost:3000
    description: Local development
paths:
...
  /warehouses:
    get:
      operationId: listWarehouses
      summary: List warehouses for the authenticated tenant
      responses:
        "200":
          description: Warehouse list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WarehouseListResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WarehouseErrorResponse"
        "403":
          description: Permission denied (requires inventory.view)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WarehouseErrorResponse"
    post:
      operationId: createWarehouse
      summary: Create a warehouse for the authenticated tenant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWarehouseRequest"
      responses:
        "201":
          description: Warehouse created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Warehouse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WarehouseErrorResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WarehouseErrorResponse"
        "403":
          description: Permission denied (requires inventory.manage)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WarehouseErrorResponse"
        "409":
          description: Warehouse code already exists for the tenant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WarehouseErrorResponse"
        "422":
          description: Free plan warehouse limit reached
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WarehouseErrorResponse"
...
  /titles:
    get:
      operationId: listTitles
      summary: List available titles for user assignment
      responses:
        "200":
          description: List of available titles
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TitleListResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserErrorResponse"
        "403":
          description: Permission denied (requires users.manage)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserErrorResponse"
  /users:
    get:
      operationId: listUsers
      summary: List tenant users with their assigned titles
      responses:
        "200":
          description: List of tenant users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserErrorResponse"
        "403":
          description: Permission denied (requires users.manage)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserErrorResponse"
    post:
      operationId: createUser
      summary: Create a new user and assign a title
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserItem"
        "400":
          description: Invalid request payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserErrorResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserErrorResponse"
        "403":
          description: Permission denied (requires users.manage)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserErrorResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserErrorResponse"
components:
  schemas:
...
    Warehouse:
      type: object
      additionalProperties: false
      required: [id, code, name, createdAt]
      properties:
        id: { type: string }
        code: { type: string }
        name: { type: string }
        createdAt: { type: string, format: date-time }
    WarehouseListResponse:
      type: object
      additionalProperties: false
      required: [items, total]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Warehouse" }
        total: { type: integer }
    CreateWarehouseRequest:
      type: object
      additionalProperties: false
      required: [code, name]
      properties:
        code: { type: string, minLength: 1, maxLength: 30 }
        name: { type: string, minLength: 1, maxLength: 120 }
    WarehouseErrorResponse:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code:
          type: string
          enum:
            [
              UNAUTHORIZED,
              FORBIDDEN,
              VALIDATION_ERROR,
              WAREHOUSE_LIMIT_REACHED,
              WAREHOUSE_CODE_EXISTS,
            ]
        message: { type: string }
...
    TitleItem:
      type: object
      additionalProperties: false
      required: [id, code, name]
      properties:
        id: { type: string }
        code: { type: string }
        name: { type: string }
    TitleListResponse:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/TitleItem"
    UserItem:
      type: object
      additionalProperties: false
      required: [id, email, fullName, status, titles, createdAt]
      properties:
        id: { type: string }
        email: { type: string, format: email }
        fullName: { type: string }
        status: { type: string, enum: [active, inactive] }
        titles:
          type: array
          items: { type: string }
        createdAt: { type: string, format: date-time }
    UserListResponse:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/UserItem"
    CreateUserRequest:
      type: object
      additionalProperties: false
      required: [email, fullName, password, titleId]
      properties:
        email:
          type: string
          format: email
        fullName:
          type: string
          minLength: 1
          maxLength: 100
        password:
          type: string
          minLength: 6
          maxLength: 100
        titleId:
          type: string
    UserErrorResponse:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code:
          type: string
          enum:
            [
              UNAUTHORIZED,
              FORBIDDEN,
              VALIDATION_ERROR,
              TITLE_NOT_FOUND,
              EMAIL_EXISTS,
              INVALID_INPUT,
            ]
        message: { type: string }